+++ SukiUI.Demo/Features/Playground/PlaygroundView.axaml.cs
@@ -18,171 +18,172 @@ using SukiUI.Controls;
 using TextMateSharp.Grammars;
 using SukiUI.Demo.Utilities;
 
-namespace SukiUI.Demo.Features.Playground
+namespace SukiUI.Demo.Features.Playground;
+
+// There are lots of null suppress warnings in this file.
+// This is because the controls are initialized in the OnApplyTemplate method.
+// The controls are guaranteed to be initialized before they are used in the methods.
+
+public partial class PlaygroundView : UserControl
 {
-    public partial class PlaygroundView : UserControl
+    private CompletionWindow? _completionWindow;
+
+    private TextEditor? _textEditor;
+
+    private GlassCard? _glassPlayground;
+
+    private Button? _renderButton;
+
+    private Button? _clearButton;
+
+    public PlaygroundView()
     {
-        private CompletionWindow? _completionWindow;
+        InitializeComponent();
+    }
 
-        private TextEditor _textEditor;
+    private void InitializeComponent()
+    {
+        AvaloniaXamlLoader.Load(this);
+    }
 
-        private GlassCard _glassPlayground;
+    protected override void OnApplyTemplate(TemplateAppliedEventArgs e)
+    {
+        base.OnApplyTemplate(e);
 
-        private Button _renderButton;
+        _textEditor = this.FindControl<TextEditor>("Editor")!;
+        _glassPlayground = this.FindControl<GlassCard>("GlassExample")!;
+        _textEditor.TextArea.TextEntered += TextEditor_TextArea_TextEntered!;
+        _textEditor.TextArea.TextEntering += TextEditor_TextArea_TextEntering!;
+        _textEditor.Text = XamlData.PlaygroundStartingCode;
+        _textEditor.TextArea.IndentationStrategy = new CSharpIndentationStrategy(_textEditor.Options);
+        _textEditor.TextArea.RightClickMovesCaret = true;
 
-        private Button _clearButton;
+        _renderButton = this.FindControl<Button>("RenderButton")!;
+        _renderButton.Click += OnRenderClicked;
+        _clearButton = this.FindControl<Button>("ClearButton")!;
+        //  _clearButton.Click += OnClearClicked;
 
-        public PlaygroundView()
+        OnBaseThemeChanged(Application.Current!.ActualThemeVariant);
+        SukiTheme.GetInstance().OnBaseThemeChanged += OnBaseThemeChanged;
+    }
+    
+    private void OnClearClicked(object? sender, RoutedEventArgs e)
+    {
+        _textEditor!.Text = string.Empty;
+        _glassPlayground!.Content = null;
+    }
+
+    private void OnRenderClicked(object? sender, RoutedEventArgs e)
+    {
+        var previewCode = XamlData.InsertIntoGridControl(_textEditor!.Text);
+
+        try
         {
-            InitializeComponent();
+            Control demoContent = AvaloniaRuntimeXamlLoader.Parse<Grid>(previewCode);
+            _glassPlayground!.Content = demoContent;
         }
-
-        private void InitializeComponent()
+        catch (XmlException ex)
         {
-            AvaloniaXamlLoader.Load(this);
+            SukiHost.ShowToast("Error", $"Exception occurred during parsing xml: \n {ex.Message}");
         }
-
-        protected override void OnApplyTemplate(TemplateAppliedEventArgs e)
+        catch (InvalidCastException ex)
         {
-            base.OnApplyTemplate(e);
-
-            _textEditor = this.FindControl<TextEditor>("Editor");
-            _glassPlayground = this.FindControl<GlassCard>("GlassExample");
-            _textEditor.TextArea.TextEntered += TextEditor_TextArea_TextEntered;
-            _textEditor.TextArea.TextEntering += TextEditor_TextArea_TextEntering;
-            _textEditor.Text = XamlData.PlaygroundStartingCode;
-            _textEditor.TextArea.IndentationStrategy = new CSharpIndentationStrategy(_textEditor.Options);
-            _textEditor.TextArea.RightClickMovesCaret = true;
-
-            _renderButton = this.FindControl<Button>("RenderButton");
-            _renderButton.Click += OnRenderClicked;
-            _clearButton = this.FindControl<Button>("ClearButton");
-          //  _clearButton.Click += OnClearClicked;
-
-            OnBaseThemeChanged(Application.Current.ActualThemeVariant);
-            SukiTheme.GetInstance().OnBaseThemeChanged += OnBaseThemeChanged;
+            SukiHost.ShowToast("Error", $"Exception occurred during conversion from xaml string to control: \n {ex.Message}");
         }
-
-        private void OnClearClicked(object? sender, RoutedEventArgs e)
+        catch (XamlLoadException ex)
         {
-            _textEditor.Text = string.Empty;
-            _glassPlayground.Content = null;
-        }
-
-        private void OnRenderClicked(object? sender, RoutedEventArgs e)
-        {
-            string previewCode = XamlData.InsertIntoGridControl(_textEditor.Text);
-
-            try
-            {
-                Control demoContent = AvaloniaRuntimeXamlLoader.Parse<Grid>(previewCode);
-                _glassPlayground.Content = demoContent;
-            }
-            catch (XmlException ex)
-            {
-                SukiHost.ShowToast("Error", $"Exception occurred during parsing xml: \n {ex.Message}");
-            }
-            catch (InvalidCastException ex)
-            {
-                SukiHost.ShowToast("Error", $"Exception occurred during conversion from xaml string to control: \n {ex.Message}");
-            }
-            catch (XamlLoadException ex)
-            {
-                SukiHost.ShowToast("Error", $"Exception occurred during loading xaml code for control: \n {ex.Message}");
-            }
-        }
-
-        private void OnBaseThemeChanged(ThemeVariant currentTheme)
-        {
-            var registryOptions = new RegistryOptions(
-                currentTheme == ThemeVariant.Dark ? ThemeName.DarkPlus : ThemeName.LightPlus);
-
-            var textMateInstallation = _textEditor.InstallTextMate(registryOptions);
-            textMateInstallation.SetGrammar(registryOptions.GetScopeByLanguageId(registryOptions
-                .GetLanguageByExtension(".xaml").Id));
-        }
-
-        private void TextEditor_TextArea_TextEntering(object sender, TextInputEventArgs e)
-        {
-            if (e.Text.Length > 0 && _completionWindow != null)
-            {
-                if (!char.IsLetterOrDigit(e.Text[0]) && _completionWindow.CompletionList.SelectedItem != null)
-                {
-                    // Whenever a non-letter is typed while the completion window is open,
-                    // insert the currently selected element.
-                    _completionWindow.CompletionList.RequestInsertion(e);
-
-                    // if space (char 32 ' ') is used for confirmation
-                    if (char.IsWhiteSpace(e.Text[0]))
-                        e.Handled = true;
-                }
-            }
-
-            // Do not set e.Handled=true.
-            // We still want to insert the character that was typed.
-        }
-
-        private void TextEditor_TextArea_TextEntered(object sender, TextInputEventArgs e)
-        {
-            if (e.Text is not ("<" or "/")) return;
-
-            _completionWindow = new CompletionWindow(_textEditor.TextArea);
-            _completionWindow.Closed += (_, _) => _completionWindow = null;
-
-            var data = _completionWindow.CompletionList.CompletionData;
-            data.Add(new MyCompletionData("suki:GlassCard"));
-            data.Add(new MyCompletionData("Grid"));
-            data.Add(new MyCompletionData("Button"));
-            data.Add(new MyCompletionData("TextBlock"));
-
-            _completionWindow.Show();
-        }
-
-
-        private void OpenPane(object? sender, RoutedEventArgs e)
-        {
-            this.Get<SplitView>("Playground").IsPaneOpen = true;
-
-            ((Button)sender).IsHitTestVisible = false;
-            ((Button)sender).Animate<double>(OpacityProperty,1,0);
-
-            this.Get<DockPanel>("TabControls").Animate<double>(OpacityProperty,0,1);
-            this.Get<DockPanel>("TabControls").IsHitTestVisible = true;
-        }
-
-        private void ClosePane(object? sender, RoutedEventArgs e)
-        {
-            this.Get<SplitView>("Playground").IsPaneOpen = false;
-
-            this.Get<Button>("OpenPaneButton").IsHitTestVisible = true;
-            this.Get<Button>("OpenPaneButton").Animate<double>(OpacityProperty,0,1);
-
-            this.Get<DockPanel>("TabControls").Animate<double>(OpacityProperty,1,0);
-            this.Get<DockPanel>("TabControls").IsHitTestVisible = false;
-        }
-
-        private void AddNewControls(object? sender, PointerPressedEventArgs e)
-        {
-            _textEditor.Text = _textEditor.Text.Insert(_textEditor.CaretOffset, ((GlassCard)sender).Tag.ToString());
+            SukiHost.ShowToast("Error", $"Exception occurred during loading xaml code for control: \n {ex.Message}");
         }
     }
 
-    internal sealed class MyCompletionData(string text) : ICompletionData
+    private void OnBaseThemeChanged(ThemeVariant currentTheme)
     {
-        public IImage? Image => null;
+        var registryOptions = new RegistryOptions(
+            currentTheme == ThemeVariant.Dark ? ThemeName.DarkPlus : ThemeName.LightPlus);
 
-        public string Text { get; } = text;
+        var textMateInstallation = _textEditor.InstallTextMate(registryOptions);
+        textMateInstallation.SetGrammar(registryOptions.GetScopeByLanguageId(registryOptions
+            .GetLanguageByExtension(".xaml").Id));
+    }
 
-        // Use this property if you want to show a fancy UIElement in the list.
-        public object Content => Text;
-
-        public object Description => "Avalonia Control";
-
-        public double Priority => 0;
-
-        public void Complete(TextArea textArea, ISegment completionSegment, EventArgs insertionRequestEventArgs)
+    private void TextEditor_TextArea_TextEntering(object sender, TextInputEventArgs e)
+    {
+        if (e.Text!.Length <= 0 || _completionWindow == null || char.IsLetterOrDigit(e.Text[0]) || _completionWindow.CompletionList.SelectedItem == null)
         {
-            textArea.Document.Replace(completionSegment, Text);
+            return;
         }
+
+        // Whenever a non-letter is typed while the completion window is open,
+        // insert the currently selected element.
+        _completionWindow.CompletionList.RequestInsertion(e);
+
+        // if space (char 32 ' ') is used for confirmation
+        if (char.IsWhiteSpace(e.Text[0]))
+            e.Handled = true;
+
+        // Do not set e.Handled=true.
+        // We still want to insert the character that was typed.
+    }
+
+    private void TextEditor_TextArea_TextEntered(object sender, TextInputEventArgs e)
+    {
+        if (e.Text is not ("<" or "/")) return;
+
+        _completionWindow = new CompletionWindow(_textEditor!.TextArea);
+        _completionWindow.Closed += (_, _) => _completionWindow = null;
+
+        var data = _completionWindow.CompletionList.CompletionData;
+        data.Add(new MyCompletionData("suki:GlassCard"));
+        data.Add(new MyCompletionData("Grid"));
+        data.Add(new MyCompletionData("Button"));
+        data.Add(new MyCompletionData("TextBlock"));
+
+        _completionWindow.Show();
+    }
+    
+    private void OpenPane(object? sender, RoutedEventArgs e)
+    {
+        this.Get<SplitView>("Playground").IsPaneOpen = true;
+
+        ((Button)sender!).IsHitTestVisible = false;
+        ((Button)sender).Animate<double>(OpacityProperty,1,0);
+
+        this.Get<DockPanel>("TabControls").Animate<double>(OpacityProperty,0,1);
+        this.Get<DockPanel>("TabControls").IsHitTestVisible = true;
+    }
+
+    private void ClosePane(object? sender, RoutedEventArgs e)
+    {
+        this.Get<SplitView>("Playground").IsPaneOpen = false;
+
+        this.Get<Button>("OpenPaneButton").IsHitTestVisible = true;
+        this.Get<Button>("OpenPaneButton").Animate<double>(OpacityProperty,0,1);
+
+        this.Get<DockPanel>("TabControls").Animate<double>(OpacityProperty,1,0);
+        this.Get<DockPanel>("TabControls").IsHitTestVisible = false;
+    }
+
+    private void AddNewControls(object? sender, PointerPressedEventArgs e)
+    {
+        _textEditor!.Text = _textEditor.Text.Insert(_textEditor.CaretOffset, ((GlassCard)sender!).Tag!.ToString()!);
+    }
+}
+
+internal sealed class MyCompletionData(string text) : ICompletionData
+{
+    public IImage? Image => null;
+
+    public string Text { get; } = text;
+
+    // Use this property if you want to show a fancy UIElement in the list.
+    public object Content => Text;
+
+    public object Description => "Avalonia Control";
+
+    public double Priority => 0;
+
+    public void Complete(TextArea textArea, ISegment completionSegment, EventArgs insertionRequestEventArgs)
+    {
+        textArea.Document.Replace(completionSegment, Text);
     }
 }
